Revert "Migrate appealable issues (#26161)"
This reverts commit b4624ac858f2f98c344abf1927d6fc021074d706.
