+++ config/settings.yml
@@ -300,14 +300,6 @@ covid_vaccine:
 database_url: <%= ENV['database_url'] %>
 decision_review:
   api_key: <%= ENV['decision_review__api_key'] %>
-  appealable_issues:
-    auth:
-      aud_claim_url: <%= ENV['decision_review__appealable_issues__auth__aud_claim_url'] %>
-      client_id: <%= ENV['decision_review__appealable_issues__auth__client_id'] %>
-      rsa_key: /srv/vets-api/secret/decision_review_appealable_issues.key
-      token_url: <%= ENV['decision_review__appealable_issues__auth__token_url'] %>
-    mock: <%= ENV['decision_review__appealable_issues__mock'] %>
-    url: <%= ENV['decision_review__appealable_issues__url'] %>
   benchmark_performance: <%= ENV['decision_review__benchmark_performance'] %>
   delete_saved_claim_records_job_enabled: <%= ENV['decision_review__delete_saved_claim_records_job_enabled'] %>
   email_pdf_upload_cutoff_date: <%= ENV['decision_review__email_pdf_upload_cutoff_date'] %>
