+++ config/settings/test.yml
@@ -299,14 +299,6 @@ covid_vaccine:
 database_url: postgis:///vets-api
 decision_review:
   api_key: abcd1234abcd1234abcd1234abcd1234abcd1234
-  appealable_issues:
-    auth:
-      aud_claim_url: ~
-      client_id: ~
-      rsa_key: ~
-      token_url: ~
-    mock: false
-    url: https://sandbox-api.va.gov/services/appeals/appealable-issues/v0
   benchmark_performance: true
   delete_saved_claim_records_job_enabled: true
   email_pdf_upload_cutoff_date: ~
