+++ modules/decision_reviews/app/controllers/decision_reviews/v1/appeals_base_controller.rb
@@ -2,7 +2,6 @@
 
 require 'caseflow/service'
 require 'decision_reviews/v1/service'
-require 'decision_reviews/v1/appealable_issues/service'
 
 module DecisionReviews
   module V1
@@ -14,10 +13,6 @@ module DecisionReviews
 
       def decision_review_service
         DecisionReviews::V1::Service.new
-      end
-
-      def appealable_issues_service
-        DecisionReviews::V1::AppealableIssues::Service.new
       end
 
       def request_body_hash
@@ -47,10 +42,6 @@ module DecisionReviews
           request_body_class_name: request.try(:body).class.name,
           request_body_string: request.try(:body).try(:string)
         }
-      end
-
-      def use_new_appealable_issues_service?
-        Flipper.enabled?(:decision_review_use_new_appealable_issues_service)
       end
     end
   end
