--- modules/decision_reviews/lib/decision_reviews/v1/appealable_issues/appealable_issues.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-require 'decision_reviews/engine'
-
-module DecisionReviews
-  module V1
-    module AppealableIssues
-    end
-  end
-end
