--- spec/support/vcr_cassettes/decision_review/SC-GET-LEGACY_APPEALS-RESPONSE-200_V1.yml
@@ -1,54 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: "<LIGHTHOUSE_CLAIMS_API_HOST>/services/appeals/v2/decision_reviews/legacy_appeals"
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept:
-      - application/json
-      Content-Type:
-      - application/json
-      User-Agent:
-      - Vets.gov Agent
-      Apikey:
-      - "<LIGHTHOUSE_API_KEY>"
-      X-Va-Ssn:
-      - '796111863'
-      X-Va-Icn:
-      - 1012832025V743496
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-  response:
-    status:
-      code: 401
-      message: Unauthorized
-    headers:
-      Date:
-      - Mon, 26 Jan 2026 21:25:28 GMT
-      Content-Type:
-      - application/json; charset=utf-8
-      Content-Length:
-      - '52'
-      Connection:
-      - keep-alive
-      Access-Control-Allow-Origin:
-      - "*"
-      Correlation-Id:
-      - 7f0a0e21-a4cf-48b9-b5b5-1a30e2b08d96#2758468
-      Cache-Control:
-      - ''
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      X-Frame-Options:
-      - SAMEORIGIN
-    body:
-      encoding: UTF-8
-      string: |-
-        {
-          "message":"Invalid authentication credentials"
-        }
-  recorded_at: Fri, 23 Jan 2026 00:00:00 GMT
-recorded_with: VCR 6.4.0
