--- spec/support/vcr_cassettes/decision_review/appealable_issues/HLR-GET-APPEALABLE-ISSUES-RESPONSE-200-MALFORMED.yml
@@ -1,80 +1,0 @@
-# Fake Bearer token for testing purposes only
----
-http_interactions:
-- request:
-    method: get
-    uri: https://sandbox-api.va.gov/services/appeals/appealable-issues/v0/appealable-issues/higher-level-reviews?benefitType=compensation&icn=1012832025V743496&receiptDate=2026-01-23
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept:
-      - application/json
-      Content-Type:
-      - application/json
-      User-Agent:
-      - Vets.gov Agent
-      Authorization: Bearer <TOKEN>
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Fri, 23 Jan 2026 21:00:38 GMT
-      Content-Type:
-      - application/json; charset=utf-8
-      Content-Length:
-      - '1200'
-      Connection:
-      - keep-alive
-      X-Ratelimit-Limit-Minute:
-      - '60'
-      Ratelimit-Remaining:
-      - '59'
-      Ratelimit-Reset:
-      - '22'
-      Ratelimit-Limit:
-      - '60'
-      X-Ratelimit-Remaining-Minute:
-      - '59'
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - '0'
-      X-Content-Type-Options:
-      - nosniff
-      X-Download-Options:
-      - noopen
-      X-Permitted-Cross-Domain-Policies:
-      - none
-      Referrer-Policy:
-      - strict-origin-when-cross-origin
-      X-Git-Sha:
-      - 48465fac678a5e09c2b3d95d8354082ad1c5ad7a
-      X-Github-Repository:
-      - https://github.com/department-of-veterans-affairs/lighthouse-vets-api
-      Content-Language:
-      - en-US
-      Vary:
-      - Accept, Origin
-      X-Request-Id:
-      - 385ed2c2-d20e-4c04-824b-d4a77faa5485
-      X-Runtime:
-      - '0.052155'
-      X-Envoy-Upstream-Service-Time:
-      - '60'
-      Strict-Transport-Security:
-      - max-age=16000000; includeSubDomains; preload;
-      - max-age=31536000; includeSubDomains; preload
-      Access-Control-Allow-Origin:
-      - "*"
-      Correlation-Id:
-      - 4839def6-ea95-4c4c-b528-e9e61971d383#2438868
-    body:
-      encoding: UTF-8
-      string: '{"data":[{"id":"a","type":"contestableIssue","attributes":{"ratingIssueReferenceId":"6160","ratingIssueProfileDate":"2005-02-14","ratingIssueDiagnosticCode":"9400","ratingIssueSubjectText":"psychosis","ratingIssuePercentNumber":"30","description":"Service connection for psychosis","isRating":true,"latestIssuesInChain":[{"id":true,"approxDecisionDate":"2005-08-22"}],"decisionIssueId":1,"ratingDecisionReferenceId":"a","approxDecisionDate":"2005-08-22","timely":false}}]}'
-  recorded_at: Fri, 23 Jan 2026 21:00:38 GMT
-recorded_with: VCR 6.4.0
