--- spec/support/vcr_cassettes/decision_review/appealable_issues/HLR-GET-APPEALABLE-ISSUES-RESPONSE-200.yml
@@ -1,103 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://sandbox-api.va.gov/services/appeals/appealable-issues/v0/appealable-issues/higher-level-reviews?benefitType=compensation&icn=1012832025V743496&receiptDate=2026-01-23
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      # Fake Bearer token for testing purposes only
-      Authorization:
-      - Bearer fake-token-12345
-      Accept:
-      - application/json
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - application/json
-    body:
-      encoding: UTF-8
-      string: |
-        {
-          "data": [
-            {
-              "id": null,
-              "type": "appealableIssue",
-              "attributes": {
-                "ratingIssueSubjectText": "Migraine",
-                "description": "Service connection for Migraine is granted with an evaluation of 50 percent effective February 1, 2020.",
-                "approxDecisionDate": "2020-02-01",
-                "decisionIssueId": 1,
-                "ratingIssuePercentNumber": "50",
-                "ratingDecisionReferenceId": "reference-id-1",
-                "latestIssuesInChain": [
-                  {
-                    "id": null,
-                    "approxDecisionDate": "2020-02-01"
-                  }
-                ]
-              }
-            },
-            {
-              "id": null,
-              "type": "appealableIssue",
-              "attributes": {
-                "ratingIssueSubjectText": "Paralysis of sciatic nerve",
-                "description": "Service connection for Paralysis of sciatic nerve is granted with an evaluation of 20 percent effective April 1, 2020.",
-                "approxDecisionDate": "2020-04-01",
-                "decisionIssueId": 2,
-                "ratingIssuePercentNumber": "20",
-                "ratingDecisionReferenceId": "reference-id-2",
-                "latestIssuesInChain": [
-                  {
-                    "id": null,
-                    "approxDecisionDate": "2020-04-01"
-                  }
-                ]
-              }
-            },
-            {
-              "id": null,
-              "type": "appealableIssue",
-              "attributes": {
-                "ratingIssueSubjectText": "PTSD",
-                "description": "Service connection for PTSD is granted with an evaluation of 70 percent effective May 1, 2020.",
-                "approxDecisionDate": "2020-05-01",
-                "decisionIssueId": 3,
-                "ratingIssuePercentNumber": "70",
-                "ratingDecisionReferenceId": "reference-id-3",
-                "latestIssuesInChain": [
-                  {
-                    "id": null,
-                    "approxDecisionDate": "2020-05-01"
-                  }
-                ]
-              }
-            },
-            {
-              "id": null,
-              "type": "appealableIssue",
-              "attributes": {
-                "ratingIssueSubjectText": "Ankle limitation",
-                "description": "Service connection for Ankle limitation is granted with an evaluation of 10 percent effective June 1, 2020.",
-                "approxDecisionDate": "2020-06-01",
-                "decisionIssueId": 4,
-                "ratingIssuePercentNumber": "10",
-                "ratingDecisionReferenceId": "reference-id-4",
-                "latestIssuesInChain": [
-                  {
-                    "id": null,
-                    "approxDecisionDate": "2020-06-01"
-                  }
-                ]
-              }
-            }
-          ]
-        }
-    http_version: null
-  recorded_at: Wed, 22 Jan 2026 00:00:00 GMT
-recorded_with: VCR 6.4.0
