--- spec/support/vcr_cassettes/decision_review/appealable_issues/NOD-GET-APPEALABLE-ISSUES-RESPONSE-422.yml
@@ -1,81 +1,0 @@
-# Fake Bearer token for testing purposes only
----
-http_interactions:
-- request:
-    method: get
-    uri: https://sandbox-api.va.gov/services/appeals/appealable-issues/v0/appealable-issues/notice-of-disagreements?benefitType=apricot&icn=2331123616V514303&receiptDate=2026-01-23
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept:
-      - application/json
-      Content-Type:
-      - application/json
-      User-Agent:
-      - Vets.gov Agent
-      Authorization: Bearer <TOKEN>
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-  response:
-    status:
-      code: 422
-      message: Unprocessable Entity
-    headers:
-      Date:
-      - Fri, 23 Jan 2026 21:00:38 GMT
-      Content-Type:
-      - application/json; charset=utf-8
-      Content-Length:
-      - '200'
-      Connection:
-      - keep-alive
-      X-Ratelimit-Limit-Minute:
-      - '60'
-      Ratelimit-Remaining:
-      - '59'
-      Ratelimit-Reset:
-      - '22'
-      Ratelimit-Limit:
-      - '60'
-      X-Ratelimit-Remaining-Minute:
-      - '59'
-      X-Frame-Options:
-      - SAMEORIGIN
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - '0'
-      X-Content-Type-Options:
-      - nosniff
-      X-Download-Options:
-      - noopen
-      X-Permitted-Cross-Domain-Policies:
-      - none
-      Referrer-Policy:
-      - strict-origin-when-cross-origin
-      X-Git-Sha:
-      - 48465fac678a5e09c2b3d95d8354082ad1c5ad7a
-      X-Github-Repository:
-      - https://github.com/department-of-veterans-affairs/lighthouse-vets-api
-      Content-Language:
-      - en-US
-      Vary:
-      - Accept, Origin
-      X-Request-Id:
-      - 385ed2c2-d20e-4c04-824b-d4a77faa5485
-      X-Runtime:
-      - '0.052155'
-      X-Envoy-Upstream-Service-Time:
-      - '60'
-      Strict-Transport-Security:
-      - max-age=16000000; includeSubDomains; preload;
-      - max-age=31536000; includeSubDomains; preload
-      Access-Control-Allow-Origin:
-      - "*"
-      Correlation-Id:
-      - 4839def6-ea95-4c4c-b528-e9e61971d383#2438868
-    body:
-      encoding: UTF-8
-      string: '{"errors":[{"status":"422","title":"Invalid benefit type","detail":"benefitType must be one of: compensation, pension, fiduciary, education, vha, voc_rehab, loan_guaranty, insurance, vre"}]}'
-  recorded_at: Fri, 23 Jan 2026 21:00:38 GMT
-recorded_with: VCR 6.4.0
