--- spec/support/vcr_cassettes/decision_review/appealable_issues/SC-GET-APPEALABLE-ISSUES-RESPONSE-200.yml
@@ -1,74 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://sandbox-api.va.gov/services/appeals/appealable-issues/v0/appealable-issues/supplemental-claims?benefitType=compensation&icn=1012832025V743496&receiptDate=2026-01-23
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept:
-      - application/json
-      Content-Type:
-      - application/json
-      User-Agent:
-      - Vets.gov Agent
-      # Fake Bearer token for testing purposes only
-      Authorization:
-      - Bearer fake-token-12345
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Wed, 22 Jan 2026 20:15:30 GMT
-      Content-Type:
-      - application/json; charset=utf-8
-      Connection:
-      - keep-alive
-      X-Ratelimit-Remaining-Minute:
-      - '59'
-      X-Ratelimit-Limit-Minute:
-      - '60'
-      Ratelimit-Remaining:
-      - '59'
-      Ratelimit-Limit:
-      - '60'
-      Ratelimit-Reset:
-      - '5'
-      Etag:
-      - W/"a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6"
-      Referrer-Policy:
-      - strict-origin-when-cross-origin
-      Vary:
-      - Origin
-      X-Content-Type-Options:
-      - nosniff
-      X-Request-Id:
-      - abc123-def456-ghi789-jkl012
-      X-Runtime:
-      - '0.012345'
-      Access-Control-Allow-Origin:
-      - "*"
-      X-Kong-Upstream-Latency:
-      - '18'
-      X-Kong-Proxy-Latency:
-      - '2'
-      Via:
-      - kong/3.2.1
-      Strict-Transport-Security:
-      - max-age=31536000; includeSubDomains; preload
-      Cache-Control:
-      - no-cache, no-store
-      Pragma:
-      - no-cache
-      Transfer-Encoding:
-      - chunked
-    body:
-      encoding: ASCII-8BIT
-      string: '{"data":[{"id":null,"type":"appealableIssue","attributes":{"ratingIssueReferenceId":"142910","ratingIssueProfileDate":"2022-03-15","ratingIssueDiagnosticCode":"8100","ratingIssueSubjectText":"Migraine","ratingIssuePercentNumber":"10","description":"Service connection for Migraine is granted with an evaluation of 10 percent effective January 15, 2020.","isRating":true,"latestIssuesInChain":[{"id":null,"approxDecisionDate":"2022-03-15"}],"decisionIssueId":null,"ratingDecisionReferenceId":null,"approxDecisionDate":"2022-03-15","rampClaimId":null,"titleOfActiveReview":null,"sourceReviewType":null,"timely":true}},{"id":null,"type":"appealableIssue","attributes":{"ratingIssueReferenceId":"142911","ratingIssueProfileDate":"2022-03-12","ratingIssueDiagnosticCode":"8520","ratingIssueSubjectText":"Paralysis of sciatic nerve","ratingIssuePercentNumber":"20","description":"Service connection for Paralysis of sciatic nerve is granted with an evaluation of 20 percent effective November 3, 2019.","isRating":true,"latestIssuesInChain":[{"id":null,"approxDecisionDate":"2022-03-12"}],"decisionIssueId":null,"ratingDecisionReferenceId":null,"approxDecisionDate":"2022-03-12","rampClaimId":null,"titleOfActiveReview":null,"sourceReviewType":null,"timely":true}},{"id":null,"type":"appealableIssue","attributes":{"ratingIssueReferenceId":"142912","ratingIssueProfileDate":"2022-02-22","ratingIssueDiagnosticCode":"9411","ratingIssueSubjectText":"PTSD","ratingIssuePercentNumber":"40","description":"Service connection for PTSD is granted with an evaluation of 40 percent effective November 3, 2019.","isRating":true,"latestIssuesInChain":[{"id":null,"approxDecisionDate":"2022-02-22"}],"decisionIssueId":null,"ratingDecisionReferenceId":null,"approxDecisionDate":"2022-02-22","rampClaimId":null,"titleOfActiveReview":null,"sourceReviewType":null,"timely":true}},{"id":null,"type":"appealableIssue","attributes":{"ratingIssueReferenceId":"142913","ratingIssueProfileDate":"2022-03-10","ratingIssueDiagnosticCode":"5270","ratingIssueSubjectText":"Ankle, limitation of motion","ratingIssuePercentNumber":"10","description":"Service connection for Ankle, limitation of motion is granted with an evaluation of 10 percent effective January 15, 2020.","isRating":true,"latestIssuesInChain":[{"id":null,"approxDecisionDate":"2022-03-10"}],"decisionIssueId":null,"ratingDecisionReferenceId":null,"approxDecisionDate":"2022-03-10","rampClaimId":null,"titleOfActiveReview":null,"sourceReviewType":null,"timely":true}},{"id":null,"type":"appealableIssue","attributes":{"ratingIssueReferenceId":"142914","ratingIssueProfileDate":"2022-03-02","ratingIssueDiagnosticCode":"5270","ratingIssueSubjectText":"Ankylosis of knee","ratingIssuePercentNumber":"10","description":"Service connection for Ankylosis of knee is granted with an evaluation of 10 percent effective January 15, 2020.","isRating":true,"latestIssuesInChain":[{"id":null,"approxDecisionDate":"2022-03-02"}],"decisionIssueId":null,"ratingDecisionReferenceId":null,"approxDecisionDate":"2022-03-02","rampClaimId":null,"titleOfActiveReview":null,"sourceReviewType":null,"timely":true}},{"id":null,"type":"appealableIssue","attributes":{"ratingIssueReferenceId":"142915","ratingIssueProfileDate":"2022-03-01","ratingIssueDiagnosticCode":"5243","ratingIssueSubjectText":"Scars, disfiguring, head, face or neck","ratingIssuePercentNumber":"10","description":"Service connection for Scars, disfiguring, head, face or neck is granted with an evaluation of 10 percent effective January 15, 2020.","isRating":true,"latestIssuesInChain":[{"id":null,"approxDecisionDate":"2022-03-01"}],"decisionIssueId":null,"ratingDecisionReferenceId":null,"approxDecisionDate":"2022-03-01","rampClaimId":null,"titleOfActiveReview":null,"sourceReviewType":null,"timely":true}},{"id":null,"type":"appealableIssue","attributes":{"ratingIssueReferenceId":"142916","ratingIssueProfileDate":"2022-03-10","ratingIssueDiagnosticCode":"6100","ratingIssueSubjectText":"Hearing loss","ratingIssuePercentNumber":"10","description":"Service connection for Hearing loss is granted with an evaluation of 10 percent effective January 15, 2020.","isRating":true,"latestIssuesInChain":[{"id":null,"approxDecisionDate":"2022-03-10"}],"decisionIssueId":null,"ratingDecisionReferenceId":null,"approxDecisionDate":"2022-03-10","rampClaimId":null,"titleOfActiveReview":null,"sourceReviewType":null,"timely":true}}]}'
-  recorded_at: Wed, 22 Jan 2026 20:15:30 GMT
-recorded_with: VCR 6.0.0
